+++ src/app/Routes.ts
@@ -6,18 +6,17 @@ import {
   RouterModule
 } from '@angular/router';
 
-import {HomeComponent} from './Components/HomeComponent';
-import {AboutComponent} from './Components/AboutComponent';
 import {ContactComponent} from './Components/ContactComponent';
 import {ModuleWithProviders} from "@angular/core";
 import {ResultComponent} from "./Results";
 import {Results1Component} from "./Results1";
+import {PhysicsComponent, MathComponent} from "./Components/SubjectComponent";
 
 const routes: Routes = [
   {
   { path: 'contact', component: ContactComponent },
-  { path: 'mathematics', component: HomeComponent },
-  { path: 'about', component: AboutComponent },
+  { path: 'mathematics', component: MathComponent },
+  { path: 'physics', component: PhysicsComponent },
   { path: 'results/:res/:unres/:n/:sn', component: ResultComponent},
   { path: 'results1/:res1/:unres1', component: Results1Component},
   { path: '', component: ContactComponent, data:{'home': true} }
